hotfix to guard against non-exist res

+++ Deployment/RemoveDevResources.ps1
@@ -1,3 +1,5 @@
 $stackRes = (az resource list --tag stack-name=contoso-customer-service-app-service | ConvertFrom-Json)
 $devRes = $stackRes | Where-Object { $_.tags.'stack-environment' -eq 'dev' }
-az resource delete --id $devRes.id
+if ($devRes -and $devRes.Length -gt 0) {
+    az resource delete --id $devRes.id
+}
